+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed bug so no warning is shown when saving `options()` to the temp folder used by whirl (#206)
 * Fixed so Quarto is started with the right renv library paths when using renv (#215)
 * Refactored log.qmd to remove read_log.
+* Added new `with_options` argument to `run()`. This is the only options being set in the child sessions now. (#187)
 
 # whirl 0.3.1